[DEV-2393] aml check improvement
