+++ src/shared/i18n/de/mail.json
@@ -122,6 +122,15 @@
                 "line4": "Alternativ kannst du auch über die DFX-Services Webseite <br> oben rechts auf den Menüpunkt und dann auf \"KYC\" klicken <br> und die Verifizierung über den \"Weiter\" Button starten.",
                 "line5": "Solltest du die Verifizierung nicht vornehmen wollen, wird das Geld vollautomatisch nach 2 Tagen retourniert."
             },
+            "name_check_despite_birthday": {
+                "title": "Einzahlung wird geprüft",
+                "salutation": "Dine Einzahlung wird geprüft",
+                "line1": "Vielen Dank für dein Vertrauen in DFX.",
+                "line2": "Leider konnte die Transaktion nicht vollautomatisch freigegeben werden und muss nun durch einen Mitarbeiter manuell geprüft werden.",
+                "line3": "Dies dauert in der Regel 24h, maximal aber 7 Tage.",
+                "line4": "Du wirst benachrichtigt sobald die Prüfung abgeschlossen ist und deine Zahlung verarbeitet wurde.",
+                "line5": ""
+            },
             "high_risk_kyc_needed": {
                 "title": "Problem bei deiner Einzahlung",
                 "salutation": "Verifizierung (KYC) ist erforderlich, um deine ausstehende Transaktion auszuführen",
