+++ src/shared/i18n/es/mail.json
@@ -122,6 +122,15 @@
                 "line4": "Alternativamente, también puede hacer clic en el elemento de menú en la parte superior derecha de la página web de DFX-Services <br> y luego en \"KYC\" <br> e iniciar la verificación a través del botón \"Continuar\".",
                 "line5": "Si no quieres llevar a cabo la verificación, el dinero se devolverá después de 2 días."
             },
+            "name_check_despite_birthday": {
+                "title": "Depósito en cuenta corriente",
+                "salutation": "Se comprobará su depósito",
+                "line1": "Gracias por confiar en DFX.",
+                "line2": "Lamentablemente, la transacción no ha podido liberarse de forma totalmente automática y ahora debe ser comprobada manualmente por un empleado.",
+                "line3": "Suele tardar 24 horas, pero un máximo de 7 días.",
+                "line4": "Se le notificará tan pronto como se haya completado la comprobación y se haya procesado su pago.",
+                "line5": ""
+            },
             "high_risk_kyc_needed": {
                 "title": "Problema con su depósito",
                 "salutation": "Se requiere verificación (KYC) para ejecutar su transacción pendiente",
