+++ src/shared/i18n/it/mail.json
@@ -122,6 +122,15 @@
                 "line4": "In alternativa, potete anche cliccare sulla voce di menu in alto a destra del sito DFX-Services <br>e poi su \"KYC\" <br>e avviare la verifica tramite il pulsante \"Continua\".",
                 "line5": "Se non vuoi eseguire la verifica, i tuoi soldi saranno rimborsati completamente in automatico dopo 2 giorni."
             },
+            "name_check_despite_birthday": {
+                "title": "Deposito in conto corrente",
+                "salutation": "Il vostro deposito sarà controllato",
+                "line1": "Grazie per la vostra fiducia in DFX.",
+                "line2": "Purtroppo non è stato possibile sbloccare la transazione in modo completamente automatico e ora deve essere controllata manualmente da un dipendente.",
+                "line3": "Questo processo richiede solitamente 24 ore, ma un massimo di 7 giorni.",
+                "line4": "Sarete informati non appena il controllo sarà stato completato e il pagamento sarà stato elaborato.",
+                "line5": ""
+            },
             "high_risk_kyc_needed": {
                 "title": "Problemi con il deposito",
                 "salutation": "La verifica (KYC) è necessaria per eseguire la transazione in corso",
